src/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==38.0.4
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
filelock==3.10.2
idna==3.4
npx==0.1.1
numpy==1.24.1
platformdirs==3.1.1
psycopg2==2.9.5
pycparser==2.21
PyJWT==1.7.1
pytz==2022.6
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.21.0
